big cleanup - removing all non-lux architectures
